requirements.txt
mkdocs
mkdocs-gen-files
mkdocs-material
pyyaml
tabulate